Promotions

+++ controllers/menuController.js
@@ -1,7 +1,7 @@
 import nedb from 'nedb-promises';
 import { currentTime } from '../utility/timeFunction.js';
 //Skapar menu-db
-export const database = new nedb({
+const database = new nedb({
     filename: './data/menu.db',
     autoload: true
 });
@@ -29,11 +29,18 @@ export const addMenuItem = async (req, res, next) => {
         const alreadyInMenu = await database.findOne({ title: title });
 
         if (!alreadyInMenu) {
+            let id;
+            if (menu.length === 0) {
+                id = 1
+            } else {
+                id = menu[menu.length - 1].id + 1
+            }
             const newMenuItem = {
-                id: menu[menu.length - 1].id + 1,
+                id: id,
                 title: title,
                 desc: desc,
                 price: price,
+                originalPrice: price,
                 createdAt: currentTime()
             }
             await database.insert(newMenuItem);
@@ -62,9 +69,15 @@ export const modifyMenuItem = async (req, res, next) => {
     try {
         const id = parseInt(req.params.id);
 
-        const { title, desc, price } = req.body;
-
         const error = new Error();
+
+        if (isNaN(id)) {
+            error.message = "Du måste ange produktens id med siffror";
+            error.status = 401;
+            throw error;
+        }
+
+        const { title, desc, price } = req.body;
 
         const alreadyInMenu = await database.findOne({ title: title });
 
@@ -83,7 +96,7 @@ export const modifyMenuItem = async (req, res, next) => {
 
         const updateItem = await database.update(
             { id: id },
-            { $set: { title, desc, price, modiefiedAt: currentTime() } }
+            { $set: { title, desc, price, originalPrice: price, modiefiedAt: currentTime() } }
         );
 
         if (updateItem === 0) {
@@ -106,10 +119,16 @@ export const modifyMenuItem = async (req, res, next) => {
 export const deleteMenuItem = async (req, res, next) => {
     try {
         const id = parseInt(req.params.id);
+        const error = new Error();
+        if (isNaN(id)) {
+            error.message = "Du måste ange produktens id med siffror";
+            error.status = 401;
+            throw error;
+        }
 
         const removeItem = await database.remove({ id: id });
         if (removeItem === 0) {
-            const error = new Error(`Kan inte hitta en produkt med id: ${id}`)
+            error.message = `Kan inte hitta en produkt med id: ${id}`
             error.status = 404;
             throw error
